wrote edge case test works#destroy

+++ app/controllers/works_controller.rb
@@ -53,6 +53,11 @@ class WorksController < ApplicationController
   def destroy
     work = Work.find_by(id: params[:id])
 
+    unless work
+      head :not_found
+      return
+    end
+
     work.destroy
 
     redirect_to works_path
